backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
psycopg2-binary
alembic
email-validator 